To avoid misleading warning in special cases
As written in comment: Failing to detect such a confidence interval is not problematic,

But it is problematic for this warning: "Mismatch between aggregated frequencies and decimals aggregated to whole numbers"

Therefore, the check for warning has now been changed.

Special cases with no variance (no suppression) also affected this warning. This has also been fixed now.